feat(settings): restructure account settings page and add change password, email, and username components
