+++ app/pages/settings/account.vue
@@ -1,50 +1,17 @@
 <script setup lang="ts">
-import { birthDateFormat } from '#imports';
-import SettingsItem from '../playground/settingsItem.vue';
-import { useUserStore } from '~/stores/user';
-import { useI18n } from 'vue-i18n';
-const { t, locale: $i18n } = useI18n();
-definePageMeta({
-  layout: 'settings',
-});
-const userStore = useUserStore();
-const user = computed(() => userStore.user);
-const birthDate = computed(() =>
-  user.value?.birthDate ? birthDateFormat(user.value.birthDate, $i18n.value) : '',
-);
+import settings from '~/components/Settings/SettingsSection/index.vue';
+definePageMeta({ layout: false });
 </script>
+
 <template>
-  <div>
-    <div class="p-4">
-      <h1 class="text-2xl font-bold">{{ t('setting.account-information') }}</h1>
-      <p class="text-muted-foreground mt-2 text-sm">
-        {{ t('setting.account-information-details') }}
-      </p>
-    </div>
-    <div v-if="user" class="mt-3 flex flex-col gap-3 overflow-hidden">
-      <SettingsItem
-        :title="t('setting.username.username')"
-        :subtitle="`@${user.username}`"
-        to="/settings/username"
-      />
-      <SettingsItem
-        :title="t('setting.change-email.label')"
-        :subtitle="`${user.email}`"
-        to="/settings/email"
-      />
-      <SettingsItem
-        :title="t('setting.date-of-birth')"
-        :subtitle="birthDate"
-        to="/settings/profile"
-      />
-      <SettingsItem
-        :title="t('setting.change-password')"
-        to="/settings/changePasswordEditor"
-        subtitle="Change Your Password at any time"
-      />
-    </div>
-    <div v-else class="p-4">
-      <p class="text-red-500">{{ $t('setting.failed-to-load-user') }}</p>
-    </div>
-  </div>
+  <NuxtLayout name="settings">
+    <template #middle>
+      <!-- Static account settings overview component -->
+      <settings />
+    </template>
+    <template #right>
+      <!-- Child route content (index/email/username/changePasswordEditor) renders here -->
+      <NuxtPage />
+    </template>
+  </NuxtLayout>
 </template>
