+++ app/pages/settings/account/email.vue
@@ -1,9 +1,9 @@
 <script lang="ts" setup>
 import { useChangeEmailStore } from '~/stores/settings/change-email';
-import ChangeEmailForm from '../../components/emailSettings/ChangeEmailForm.vue';
-import OtpForm from '../../components/emailSettings/OtpForm.vue';
+import ChangeEmailForm from '~/components/emailSettings/ChangeEmailForm.vue';
+import OtpForm from '~/components/emailSettings/OtpForm.vue';
 definePageMeta({
-  layout: 'settings',
+  layout: false,
 });
 
 const userStore = useUserStore();
